+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Mark `FactoryBot/CreateList` as `SafeAutoCorrect: false`. ([@r7kamura])
+- Change `FactoryBot/CreateList` so that it considers `times.map`. ([@r7kamura])
 
 ## 2.22.0 (2023-05-04)