+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `FactoryBot/RedundantFactoryOption` cop. ([@r7kamura])
 - Add `ExplicitOnly` configuration option to `FactoryBot/ConsistentParenthesesStyle`, `FactoryBot/CreateList` and `FactoryBot/FactoryNameStyle`. ([@ydah])
 - Change `FactoryBot/CreateList` so that it checks same factory calls in an Array. ([@r7kamura])
+- Add `FactoryBot/AssociationStyle` cop. ([@r7kamura])
 
 ## 2.22.0 (2023-05-04)